+++ src/components/footer/constants.ts
@@ -1,0 +1,4 @@
+export const CONTROL_BUTTON_STYLES =
+  'h-[25px] w-[25px] text-secondary-800 active:bg-brand-200 hover:text-secondary-500';
+
+export const CONTROL_ICON_STYLES = 'h-5 w-5 self-baseline';